Rename Input to InputGridCell
I am pretty unsure about the appropriate name for a cell...
